implement multiple-tags behavior
Place a window in a set of groups rather than a single group.

We need to ensure that no two groups which share a window are ever shown
on a screen at the same time.  This patch doesn't provide that guarantee,
as we need to check for that in togroup() etc.

This patch switches out window.group for an array window.groups.  Only
one of the groups may have a screen (i.e. be currently shown) at any one
time.

The new 'addgroup' function adds a new group to the current window.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>